requirements.txt
fastapi
uvicorn[standard]
pydantic
requests
openai
python-dotenv
